♻️ Refactor: Session 카메라/마이크 조작 개선

+++ app/src/main/java/gdsc/solutionchallenge/saybetter/saybetterlearner/model/data/local/entity/SymbolRecord.kt
@@ -4,7 +4,7 @@ import java.sql.Timestamp
 
 
 data class SymbolRecord (
-    val symbolId:Long,
+    val symbolId:Int,
     val touchOrder:Long,
     val touchTime : Timestamp,
 )
